Always treat Target::link_libraries as libraries
If the link library is an absolute path, we obtain only the core
name part (e.g. for /usr/lib/libfoo.so it's "foo"), so it then
get's passed to linked as -lfoo. If the library is not an absolute
path, we just use the name as-is.

This also includes some new testcase that uses a custom test CMake
package that we try to find in tests and check its output.