Actually return None when failing to find a target
The documentation for find_target says that it returns None if the
target was found, but this isn't actually true.

The reason for this being that the Target struct uses the default serde
feature. so find_target will happily consider an empty JSON object (e.g.
"{}") as a valid target - since it was fine to parse.

This can easily be fixed by checking if the target name is empty, which
shouldn't happen for actually found targets.